call Pair/Pool template create
